Enum appstream::enums::ContentAttribute

#[non_exhaustive]
pub enum ContentAttribute {

    ViolenceCartoon(ContentState),
    ViolenceFantasy(ContentState),
    ViolenceRealistic(ContentState),
    ViolenceBloodshed(ContentState),
    ViolenceSexual(ContentState),
    ViolenceDesecration(ContentState),
    ViolenceSlavery(ContentState),
    ViolenceWorship(ContentState),
    DrugsAlcohol(ContentState),
    DrugsNarcotics(ContentState),
    DrugsTobacco(ContentState),
    SexNudity(ContentState),
    SexThemes(ContentState),
    SexHomosexuality(ContentState),
    SexProstitution(ContentState),
    SexAdultery(ContentState),
    SexAppearance(ContentState),
    LanguageProfanity(ContentState),
    LanguageHumor(ContentState),
    LanguageDiscrimination(ContentState),
    SocialChat(ContentState),
    SocialInfo(ContentState),
    SocialAudio(ContentState),
    SocialLocation(ContentState),
    SocialContacts(ContentState),
    MoneyAdvertising(ContentState),
    MoneyPurchasing(ContentState),
    MoneyGambling(ContentState),

}

OARS attribute.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ViolenceCartoon(ContentState)

Defined as fictional characters depicted in an animated film or a comic strip which do not look human.

Tuple Fields of ViolenceCartoon

0: ContentState

ViolenceFantasy(ContentState)

Defined as characters easily distinguishable from reality.

Tuple Fields of ViolenceFantasy

0: ContentState

ViolenceRealistic(ContentState)

Defined as characters not easily distinguishable from reality.

Tuple Fields of ViolenceRealistic

0: ContentState

ViolenceBloodshed(ContentState)

Defined as the killing or wounding of people.

Tuple Fields of ViolenceBloodshed

0: ContentState

ViolenceSexual(ContentState)

Defined as any unwanted sexual act or activity.

Tuple Fields of ViolenceSexual

0: ContentState

ViolenceDesecration(ContentState)

Defined as the action of desecrating something, typically a human body.

Tuple Fields of ViolenceDesecration

0: ContentState

ViolenceSlavery(ContentState)

Defined as working without proper remuneration or appreciation.

Tuple Fields of ViolenceSlavery

0: ContentState

ViolenceWorship(ContentState)

Defined as violence targeted to places of worship.

Tuple Fields of ViolenceWorship

0: ContentState

DrugsAlcohol(ContentState)

Defined as usage of alcohol or seeing a character consumes one.

Tuple Fields of DrugsAlcohol

0: ContentState

DrugsNarcotics(ContentState)

Defined as an addictive drug affecting mood or behaviour that is specifically illegal in at least one country.

Tuple Fields of DrugsNarcotics

0: ContentState

DrugsTobacco(ContentState)

Defined as any nicotine-rich product.

Tuple Fields of DrugsTobacco

0: ContentState

SexNudity(ContentState)

Defined as a state of undress, and in this case specifically specifically nudity likely to cause offense.

Tuple Fields of SexNudity

0: ContentState

SexThemes(ContentState)

Defined as in reference to a sexual act.

Tuple Fields of SexThemes

0: ContentState

SexHomosexuality(ContentState)

Defined as sexual attraction to people of one’s own sex.

Tuple Fields of SexHomosexuality

0: ContentState

SexProstitution(ContentState)

Defined as the practice or occupation of engaging in sexual activity with someone for payment.

Tuple Fields of SexProstitution

0: ContentState

SexAdultery(ContentState)

Defined as voluntary interaction between a married person and a person who is not their spouse.

Tuple Fields of SexAdultery

0: ContentState

SexAppearance(ContentState)

Defined as appearance of human or human-like characters that are sexualized in some way.

Tuple Fields of SexAppearance

0: ContentState

LanguageProfanity(ContentState)

Defined as blasphemous or obscene language.

Tuple Fields of LanguageProfanity

0: ContentState

LanguageHumor(ContentState)

Defined as the quality of being amusing.

Tuple Fields of LanguageHumor

0: ContentState

LanguageDiscrimination(ContentState)

Defined as the unjust or prejudicial treatment of different categories of people, especially on the grounds of race, age, or sex.

Tuple Fields of LanguageDiscrimination

0: ContentState

SocialChat(ContentState)

Defined as any messaging system connected to the Internet.

Tuple Fields of SocialChat

0: ContentState

SocialInfo(ContentState)

Defined as sharing information with a legal entity typically used for advertising or for sending back diagnostic data.

Tuple Fields of SocialInfo

0: ContentState

SocialAudio(ContentState)

Defined as any multimedia messaging system connected to the Internet.

Tuple Fields of SocialAudio

0: ContentState

SocialLocation(ContentState)

Defined as sharing your physical real-time location.

Tuple Fields of SocialLocation

0: ContentState

SocialContacts(ContentState)

Defined as sharing identifiable details with other users to allow out-of-band communication.

Tuple Fields of SocialContacts

0: ContentState

MoneyAdvertising(ContentState)

Defined as the activity of producing advertisements for commercial products or services.

Tuple Fields of MoneyAdvertising

0: ContentState

MoneyPurchasing(ContentState)

Defined as items or points that a user can buy for use within a virtual world to improve a character or enhance the playing experience.

Tuple Fields of MoneyPurchasing

0: ContentState

MoneyGambling(ContentState)

Defined as taking a risky action in the hope of a desired result.

Tuple Fields of MoneyGambling

0: ContentState