Struct appstream::Component

pub struct Component {
    pub kind: ComponentKind,
    pub id: AppId,
    pub name: TranslatableString,
    pub summary: Option<TranslatableString>,
    pub description: Option<MarkupTranslatableString>,
    pub project_license: Option<License>,
    pub metadata_license: Option<License>,
    pub project_group: Option<String>,
    pub compulsory_for_desktop: Option<String>,
    pub extends: Vec<AppId>,
    pub icons: Vec<Icon>,
    pub screenshots: Vec<Screenshot>,
    pub urls: Vec<ProjectUrl>,
    pub developer_name: Option<TranslatableString>,
    pub update_contact: Option<String>,
    pub categories: Vec<Category>,
    pub launchables: Vec<Launchable>,
    pub pkgname: Option<String>,
    pub source_pkgname: Option<String>,
    pub bundles: Vec<Bundle>,
    pub releases: Vec<Release>,
    pub languages: Vec<Language>,
    pub mimetypes: Vec<String>,
    pub kudos: Vec<Kudo>,
    pub keywords: Option<TranslatableList>,
    pub content_rating: Option<ContentRating>,
    pub provides: Vec<Provide>,
    pub translations: Vec<Translation>,
}

A component is wrapper around a metainfo.xml file or previously an appdata.xml file. It describes an application to the various stores out there on Linux.

Fields
kind: ComponentKind

The component type.

id: AppId

Unique identifier for this component.

name: TranslatableString

A human-readable name.

summary: Option<TranslatableString>

A short summary of the component.

description: Option<MarkupTranslatableString>

A long description of this component.

project_license: Option<License>

The license of the compoonent.

metadata_license: Option<License>

The license the metainfo XML file released under.

project_group: Option<String>

Identify the project with a specific upstream umbrella. Known values includes: GNOME, KDE, XFCE, MATE, LXDE.

compulsory_for_desktop: Option<String>

Indicate for which desktop environment the component is essential for its functionality.

extends: Vec<AppId>

The various AppId that the current component extends.

icons: Vec<Icon>

The icons of the component.

screenshots: Vec<Screenshot>

The component screenshots, composed of either images, videos or both.

urls: Vec<ProjectUrl>

Web URLs.

developer_name: Option<TranslatableString>

The developers or the projects responsible for the development of the project.

update_contact: Option<String>

Used by distributors to contact the project. The information should not be exposed to the user.

categories: Vec<Category>

The categories this component is associated with.

launchables: Vec<Launchable>

Possible methods to launch the software.

pkgname: Option<String>

The pkgname, a distributor thing.

source_pkgname: Option<String>

The source pkgname, a distributor thing.

bundles: Vec<Bundle>

3rd-party sources to grab the component from.

releases: Vec<Release>

Metainformation that describes the various releases.

languages: Vec<Language>

The languages supported by the component.

mimetypes: Vec<String>

The MIME types the component supports.

kudos: Vec<Kudo>

Defines the "awesomeness" of a component.

keywords: Option<TranslatableList>

A list of keywords, to help the user find the component easily.

content_rating: Option<ContentRating>

Specifies the age rating of the component.

provides: Vec<Provide>

Public interfaces the component provides.

translations: Vec<Translation>

Specifies the translation domains.

Implementations

impl Component

pub fn from_path(path: PathBuf) -> Result<Self, ParseError>

Create a new Component from an XML file.

Arguments

path - The path to the component.

pub fn from_gzipped(path: PathBuf) -> Result<Self, ParseError>

Create a new Component from a gzipped XML file.

Arguments

path - The path to the gzipped component.