Struct appstream::builders::ComponentBuilder

pub struct ComponentBuilder {
    pub kind: ComponentKind,
    pub id: Option<AppId>,
    pub name: Option<TranslatableString>,
    pub summary: Option<TranslatableString>,
    pub description: Option<MarkupTranslatableString>,
    pub project_license: Option<License>,
    pub metadata_license: Option<License>,
    pub project_group: Option<String>,
    pub compulsory_for_desktop: Option<String>,
    pub extends: Vec<AppId>,
    pub icons: Vec<Icon>,
    pub screenshots: Vec<Screenshot>,
    pub urls: Vec<ProjectUrl>,
    pub developer_name: Option<TranslatableString>,
    pub update_contact: Option<String>,
    pub categories: Vec<Category>,
    pub launchables: Vec<Launchable>,
    pub pkgname: Option<String>,
    pub bundles: Vec<Bundle>,
    pub releases: Vec<Release>,
    pub languages: Vec<Language>,
    pub mimetypes: Vec<String>,
    pub kudos: Vec<Kudo>,
    pub keywords: Option<TranslatableList>,
    pub content_rating: Option<ContentRating>,
    pub provides: Vec<Provide>,
    pub translations: Vec<Translation>,
    pub source_pkgname: Option<String>,
}

A helper to build a Component.

Fields
kind: ComponentKind

The component type.

id: Option<AppId>

A unique identifier of the component.

name: Option<TranslatableString>

The component name.

summary: Option<TranslatableString>

A short summary.

description: Option<MarkupTranslatableString>

A long description that might contains markup.

project_license: Option<License>

The project license.

metadata_license: Option<License>

The license of the metainfo.

project_group: Option<String>

The project group of the component.

compulsory_for_desktop: Option<String>

Indicate for which desktop environment the component is essential for its functionality.

extends: Vec<AppId>

The various AppId that the current component extends.

icons: Vec<Icon>

The component icons.

screenshots: Vec<Screenshot>

The component screenshots, composed of either images, videos or both.

urls: Vec<ProjectUrl>

Web URLs.

developer_name: Option<TranslatableString>

The developers or the projects responsible for the development of the project.

update_contact: Option<String>

Used by distributors to contact the project.

categories: Vec<Category>

The categories this component is associated with.

launchables: Vec<Launchable>

Possible methods to launch the software.

pkgname: Option<String>

The pkgname, a distributor thing.

bundles: Vec<Bundle>

3rd-party sources to grab the component from.

releases: Vec<Release>

Metainformation that describes the various releases.

languages: Vec<Language>

The languages supported by the component.

mimetypes: Vec<String>

The MIME types the component supports.

kudos: Vec<Kudo>

Defines the "awesomeness" of a component.

keywords: Option<TranslatableList>

A list of keywords, to help the user find the component easily.

content_rating: Option<ContentRating>

Specifies the age rating of the component.

provides: Vec<Provide>

Public interfaces the component provides.

translations: Vec<Translation>

Specifies the translation domains.

source_pkgname: Option<String>

The source pkgname, a distributor thing.

Implementations

impl ComponentBuilder

pub fn id(self, id: AppId) -> Self

Sets the component's unique identifier.

pub fn name(self, name: TranslatableString) -> Self

Sets the component name.

pub fn content_rating(self, content_rating: ContentRating) -> Self

Specifies the age rating of component.

pub fn kind(self, kind: ComponentKind) -> Self

Sets the component type.

pub fn developer_name(self, developer_name: TranslatableString) -> Self

Sets the developer name.

pub fn summary(self, summary: TranslatableString) -> Self

Sets the component summary.

pub fn description(self, description: MarkupTranslatableString) -> Self

Sets the component description.

pub fn metadata_license(self, license: License) -> Self

Sets the metainfo license.

pub fn project_license(self, license: License) -> Self

Sets the project license.

pub fn keywords(self, keywords: TranslatableList) -> Self

Sets the keywords.

pub fn compulsory_for_desktop(self, compulsory_for_desktop: &str) -> Self

Sets which desktop environment the component is essential for its functionality.

pub fn project_group(self, group: &str) -> Self

Sets the upstream umberall. Known values includes: GNOME, KDE, XFCE, MATE, LXDE.

pub fn url(self, url: ProjectUrl) -> Self

Adds a Web URL to the component.

pub fn screenshot(self, screenshot: Screenshot) -> Self

Adds a screenshot to the component.

pub fn icon(self, icon: Icon) -> Self

Adds an icon to the component.

pub fn kudo(self, kudo: Kudo) -> Self

Adds a kudo to the component.

pub fn translation(self, translation: Translation) -> Self

Adds a translation context to the component.

pub fn bundle(self, bundle: Bundle) -> Self

Adds a bundle to the component.

pub fn language(self, language: Language) -> Self

Adds a language to the component.

pub fn category(self, category: Category) -> Self

Adds a category to the component.

pub fn mimetype(self, mimetype: &str) -> Self

Adds a mimetype to the component.

pub fn extend(self, extend: AppId) -> Self

Adds a component that the current one extends.

pub fn release(self, release: Release) -> Self

Adds a release to the component.

pub fn launchable(self, launchable: Launchable) -> Self

Adds a launchable to the component.

pub fn provide(self, provide: Provide) -> Self

Adds a provided interface to the component.

pub fn pkgname(self, pkgname: &str) -> Self

Sets the pkgname, a distributor thing.

pub fn source_pkgname(self, source_pkgname: &str) -> Self

Sets the source pkgname, a distributor thing.

pub fn update_contact(self, update_contact: &str) -> Self

Sets a way to contact the developer of the project.

pub fn build(self) -> Component

Constructs a Component.