Module appstream::enums

Various enumerations used in the appstream types.

Enums

ArtifactKind

The artifact type.

Bundle

Indicates that the software is available via a 3rd-party application installer. See <bundle/>.

Category

Specifies a number of defined categories a component can be assigned to. See the list of the Registered Categories.

Checksum

Defines a checksum to validate the integrity of an artifact.

ComponentKind

Defines the various types of a Component.

ContentAttribute

OARS attribute.

ContentRatingVersion

Defines the version of the OARS specification.

ContentState

Defines the state of a ContentAttribute

FirmwareKind

Defines the firmware type.

Icon

Defines a component icon. See <icon/>.

ImageKind

The type of an image.

Kudo

Defines some metrics of awesomeness. See the specs for more information.

Launchable

Indicates possible methods to launch the application. See <launchable/>.

ProjectUrl

Defines a list of possible project URLs. See <url/>.

Provide

Describes the public interfaces the component provides. See <provide/>.

ReleaseKind

Classifies the release into stable/development. See <releases/>.

ReleaseUrgency

Defines how important is to install the new release as un update. See <releases/>.

Size

Defines the download and installed size of a Component or Artifact.

Translation

Defines the possible translation domains. See <translation/>.