Enum appstream::enums::Category

#[non_exhaustive]
pub enum Category {

    AudioVideo,
    Audio,
    Video,
    Development,
    Education,
    Game,
    Graphics,
    Network,
    Office,
    Science,
    Settings,
    System,
    Utility,
    Building,
    Debugger,
    IDE,
    GUIDesigner,
    Profiling,
    RevisionControl,
    Translation,
    Calendar,
    ContactManagement,
    Database,
    Dictionary,
    Chart,
    Email,
    Finance,
    FlowChart,
    PDA,
    ProjectManagement,
    Presentation,
    Spreadsheet,
    WordProcessor,
    TwoDGraphics,
    VectorGraphics,
    RasterGraphics,
    ThreeDGraphics,
    Scanning,
    OCR,
    Photography,
    Publishing,
    Viewer,
    TextTools,
    DesktopSettings,
    HardwareSettings,
    Printing,
    PackageManager,
    Dialup,
    InstantMessaging,
    Chat,
    IRCClient,
    Feed,
    FileTransfer,
    HamRadio,
    News,
    P2P,
    RemoteAccess,
    Telephony,
    TelephonyTools,
    VideoConference,
    WebBrowser,
    WebDevelopment,
    Midi,
    Mixer,
    Sequencer,
    Tuner,
    TV,
    AudioVideoEditing,
    Player,
    Recorder,
    DiscBurning,
    ActionGame,
    AdventureGame,
    ArcadeGame,
    BoardGame,
    BlocksGame,
    CardGame,
    KidsGame,
    LogicGame,
    RolePlaying,
    Shooter,
    Simulation,
    SportsGame,
    StrategyGame,
    Art,
    Construction,
    Music,
    Languages,
    ArtificialIntelligence,
    Astronomy,
    Biology,
    Chemistry,
    ComputerScience,
    DataVisualization,
    Economy,
    Electricity,
    Geography,
    Geology,
    Geoscience,
    History,
    Humanities,
    ImageProcessing,
    Literature,
    Maps,
    Math,
    NumericalAnalysis,
    MedicalSoftware,
    Physics,
    Robotics,
    Spirituality,
    Sports,
    ParallelComputing,
    Amusement,
    Archiving,
    Compression,
    Electronics,
    Emulator,
    Engineering,
    FileTools,
    FileManager,
    TerminalEmulator,
    Filesystem,
    Monitor,
    Security,
    Accessibility,
    Calculator,
    Clock,
    TextEditor,
    Documentation,
    Adult,
    Core,
    KDE,
    GNOME,
    XFCE,
    GTK,
    Qt,
    Motif,
    Java,
    ConsoleOnly,
    Screensaver,
    TrayIcon,
    Applet,
    Shell,
    // some variants omitted

}

Specifies a number of defined categories a component can be assigned to. See the list of the Registered Categories.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AudioVideo

Application for presenting, creating, or processing multimedia (audio/video).

Audio

An audio application.

Video

A video application.

Development

An application for development.

Education

Educational software.

Game

A game.

Graphics

Application for viewing, creating, or processing graphics.

Network

Network application such as a web browser.

Office

An office type application.

Science

Scientific software.

Settings

Settings applications.

System

System application, “System Tools” such as say a log viewer or network monitor.

Utility

Small utility application, “Accessories”.

Building

A tool to build applications.

Debugger

A tool to debug applications.

IDE

IDE application.

GUIDesigner

A GUI designer application.

Profiling

A profiling tool.

RevisionControl

Applications like cvs or subversion.

Translation

A translation tool.

Calendar

Calendar application.

ContactManagement

E.g. an address book.

Database

Application to manage a database.

Dictionary

A dictionary.

Chart

Chart application.

Email

Email application.

Finance

Application to manage your finance.

FlowChart

A flowchart application.

PDA

Tool to manage your PDA.

ProjectManagement

Project management application.

Presentation

Presentation software.

Spreadsheet

A spreadsheet.

WordProcessor

A word processor.

TwoDGraphics

2D based graphical application.

VectorGraphics

Application for viewing, creating, or processing vector graphics.

RasterGraphics

Application for viewing, creating, or processing raster (bitmap) graphics.

ThreeDGraphics

Application for viewing, creating, or processing 3-D graphics.

Scanning

Tool to scan a file/text.

OCR

Optical character recognition application.

Photography

Camera tools, etc.

Publishing

Desktop Publishing applications and Color Management tools.

Viewer

Tool to view e.g. a graphic or pdf file.

TextTools

A text tool utility.

DesktopSettings

Configuration tool for the GUI.

HardwareSettings

A tool to manage hardware components, like sound cards, video cards or printers.

Printing

A tool to manage printers.

PackageManager

A package manager application.

Dialup

A dial-up program.

InstantMessaging

An instant messaging client.

Chat

A chat client

IRCClient

An IRC client.

Feed

RSS, podcast and other subscription based contents.

FileTransfer

Tools like FTP or P2P programs.

HamRadio

HAM radio software.

News

A news reader or a news ticker.

P2P

A P2P program.

RemoteAccess

A tool to remotely manage your PC.

Telephony

Telephony via PC.

TelephonyTools

Telephony tools, to dial a number, manage PBX, …

VideoConference

Video Conference software.

WebBrowser

A web browser.

WebDevelopment

A tool for web developers

Midi

An app related to MIDI.

Mixer

Just a mixer.

Sequencer

A sequencer.

Tuner

A tuner.

TV

A TV application.

AudioVideoEditing

Application to edit audio/video files.

Player

Application to play audio/video files.

Recorder

Application to record audio/video files.

DiscBurning

Application to burn a disc.

ActionGame

An action game.

AdventureGame

Adventure style game.

ArcadeGame

Arcade style game.

BoardGame

A board game.

BlocksGame

Falling blocks game.

CardGame

A card game.

KidsGame

A game for kids.

LogicGame

Logic games like puzzles, etc.

RolePlaying

A role playing game.

Shooter

A shooter game.

Simulation

A simulation game.

SportsGame

A sports game.

StrategyGame

A strategy game.

Art

Software to teach arts.

Construction

Music

Musical software.

Languages

Software to learn foreign languages.

ArtificialIntelligence

Artificial Intelligence software.

Astronomy

Astronomy software.

Biology

Biology software.

Chemistry

Chemistry software.

ComputerScience

Computer sience software.

DataVisualization

Data visualization software.

Economy

Economy software.

Electricity

Electricity software.

Geography

Geography software.

Geology

Geology software.

Geoscience

Geoscience software, GIS.

History

History software.

Humanities

Software for philosophy, psychology and other humanities.

ImageProcessing

Image Processing software.

Literature

Literature software.

Maps

Software for viewing maps, navigation, mapping, GPS.

Math

Math software.

NumericalAnalysis

Numerical analysis software.

MedicalSoftware

Medical software.

Physics

Physics software.

Robotics

Robotics software.

Spirituality

Religious and spiritual software, theology.

Sports

Sports software.

ParallelComputing

Parallel computing software.

Amusement

A simple amusement.

Archiving

A tool to archive/backup data.

Compression

A tool to manage compressed data/archives.

Electronics

Electronics software, e.g. a circuit designer.

Emulator

Emulator of another platform, such as a DOS emulator.

Engineering

Engineering software, e.g. CAD programs.

FileTools

A file tool utility.

FileManager

A file manager.

TerminalEmulator

A terminal emulator application.

Filesystem

A file system tool.

Monitor

Monitor application/applet that monitors some resource or activity.

Security

A security tool.

Accessibility

Accessibility.

Calculator

A calculator.

Clock

A clock application/applet.

TextEditor

A text editor.

Documentation

Help or documentation.

Adult

Application handles adult or explicit material.

Core

Important application, core to the desktop such as a file manager or a help browser.

KDE

Application based on KDE libraries.

GNOME

Application based on GNOME libraries.

XFCE

Application based on XFCE libraries.

GTK

Application based on GTK+ libraries.

Qt

Application based on Qt libraries.

Motif

Application based on Motif libraries.

Java

Application based on Java GUI libraries, such as AWT or Swing.

ConsoleOnly

Application that only works inside a terminal (text-based or command line application).

Screensaver

A screen saver.

TrayIcon

An application that is primarily an icon for the “system tray” or “notification area”.

Applet

An applet that will run inside a panel or another such application, likely desktop specific.

Shell

A shell (an actual specific shell such as bash or tcsh, not a TerminalEmulator).