Enum appstream::enums::ContentAttribute

#[non_exhaustive]pub enum ContentAttribute {
    ViolenceCartoon(ContentState),
    ViolenceFantasy(ContentState),
    ViolenceRealistic(ContentState),
    ViolenceBloodshed(ContentState),
    ViolenceSexual(ContentState),
    ViolenceDesecration(ContentState),
    ViolenceSlavery(ContentState),
    ViolenceWorship(ContentState),
    DrugsAlcohol(ContentState),
    DrugsNarcotics(ContentState),
    DrugsTobacco(ContentState),
    SexNudity(ContentState),
    SexThemes(ContentState),
    SexHomosexuality(ContentState),
    SexProstitution(ContentState),
    SexAdultery(ContentState),
    SexAppearance(ContentState),
    LanguageProfanity(ContentState),
    LanguageHumor(ContentState),
    LanguageDiscrimination(ContentState),
    SocialChat(ContentState),
    SocialInfo(ContentState),
    SocialAudio(ContentState),
    SocialLocation(ContentState),
    SocialContacts(ContentState),
    MoneyAdvertising(ContentState),
    MoneyPurchasing(ContentState),
    MoneyGambling(ContentState),
}

OARS attribute.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ViolenceCartoon(ContentState)

Defined as fictional characters depicted in an animated film or a comic strip which do not look human.

ViolenceFantasy(ContentState)

Defined as characters easily distinguishable from reality.

ViolenceRealistic(ContentState)

Defined as characters not easily distinguishable from reality.

ViolenceBloodshed(ContentState)

Defined as the killing or wounding of people.

ViolenceSexual(ContentState)

Defined as any unwanted sexual act or activity.

ViolenceDesecration(ContentState)

Defined as the action of desecrating something, typically a human body.

ViolenceSlavery(ContentState)

Defined as working without proper remuneration or appreciation.

ViolenceWorship(ContentState)

Defined as violence targeted to places of worship.

DrugsAlcohol(ContentState)

Defined as usage of alcohol or seeing a character consumes one.

DrugsNarcotics(ContentState)

Defined as an addictive drug affecting mood or behaviour that is specifically illegal in at least one country.

DrugsTobacco(ContentState)

Defined as any nicotine-rich product.

SexNudity(ContentState)

Defined as a state of undress, and in this case specifically specifically nudity likely to cause offense.

SexThemes(ContentState)

Defined as in reference to a sexual act.

SexHomosexuality(ContentState)

Defined as sexual attraction to people of one's own sex.

SexProstitution(ContentState)

Defined as the practice or occupation of engaging in sexual activity with someone for payment.

SexAdultery(ContentState)

Defined as voluntary interaction between a married person and a person who is not their spouse.

SexAppearance(ContentState)

Defined as appearance of human or human-like characters that are sexualized in some way.

LanguageProfanity(ContentState)

Defined as blasphemous or obscene language.

LanguageHumor(ContentState)

Defined as the quality of being amusing.

LanguageDiscrimination(ContentState)

Defined as the unjust or prejudicial treatment of different categories of people, especially on the grounds of race, age, or sex.

SocialChat(ContentState)

Defined as any messaging system connected to the Internet.

SocialInfo(ContentState)

Defined as sharing information with a legal entity typically used for advertising or for sending back diagnostic data.

SocialAudio(ContentState)

Defined as any multimedia messaging system connected to the Internet.

SocialLocation(ContentState)

Defined as sharing your physical real-time location.

SocialContacts(ContentState)

Defined as sharing identifiable details with other users to allow out-of-band communication.

MoneyAdvertising(ContentState)

Defined as the activity of producing advertisements for commercial products or services.

MoneyPurchasing(ContentState)

Defined as items or points that a user can buy for use within a virtual world to improve a character or enhance the playing experience.

MoneyGambling(ContentState)

Defined as taking a risky action in the hope of a desired result.